refactor(ResourceMenu): resource menu items split into separate file and list items now rendered through map()
